--- pktvisor_exporter/.gitignore
@@ -1,10 +1,0 @@
-# Go
-pktvisor_exporter
-
-# Files
-.DS_Store
-
-# Editor swap files
-*.swp
-*.swo
-*.swn
