--- pktvisor_exporter/README.md
@@ -1,9 +1,0 @@
-### Prometheus exporter for pktvisor
-
-### Local Deployment
-
-* Launch exporter:
-
-```bash
-$ pktvisor_exporter -log.level=debug
-```
