+++ reporting/pktvisor_prometheus/.gitignore
@@ -1,0 +1,10 @@
+# Go
+pktvisor_exporter
+
+# Files
+.DS_Store
+
+# Editor swap files
+*.swp
+*.swo
+*.swn
