+++ reporting/pktvisor_prometheus/README.md
@@ -1,0 +1,9 @@
+### Prometheus exporter for pktvisor
+
+### Local Deployment
+
+* Launch exporter:
+
+```bash
+$ pktvisor_exporter -log.level=debug
+```
